Feat: Search news by query with a search bar
Implemented a new bloc for the external news to handle the two different events:
	RequestInitialExteriorNewsEvent: Which shows mexican sports news
	RequestCustomExteriorNewsEvent: Which shows custom news by a query

Added the bloc provider at home/home_page and initialized it with RequestInitialExteriorNewsEvent, so it shows the sport news first.

Refactored noticias_deportes to use a blocConsumer and adapt to the different events.

Modified news_repository so it can handle a query parameter